% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/czytaj.logi.R
\name{czytaj.logi}
\alias{czytaj.logi}
\title{Funkcja wczytująca dane z logów.}
\usage{
czytaj.logi(pliki, baza)
}
\arguments{
\item{pliki}{Scieżka do folderu z danymi, które mają zostać wczytane.}

\item{baza}{Ścieżka do bazy SQLite, do której będą zapisywane dane.}
}
\description{
Wczytuje dane z logów znajdujących się w zadanych plikach do bazy SQLite-owej.
}

