% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dane.z.pliku.R
\name{dane.z.pliku}
\alias{dane.z.pliku}
\title{Wczytywanie danych o interakcjach z pliku.}
\usage{
dane.z.pliku(plik)
}
\arguments{
\item{plik}{Ścieżka do pliku z logami cnk.}
}
\value{
data.frame Ramka zawierająca dzień, godzinę oraz id odwiedzającego.
}
\description{
Wczytuje z pliku logów informacje o interakcjach i zwraca je w postaci
ramki danych.
}

