% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dane.ze.sciezki.R
\name{dane.ze.sciezki}
\alias{dane.ze.sciezki}
\title{Funkja wczytuje nazwę eksponatu, rok i miesiąc ze ścieżki do pliku.}
\usage{
dane.ze.sciezki(plik)
}
\arguments{
\item{plik}{Scieżka do pliku z logiem cnk.}
}
\value{
list Lista zawierająca nazwę eksponatu, rok i miesiąc użycia.
}
\description{
Wczytuje nazwę eksponatu, rok i miesiąc ze ścieżki do pliku cnk.
}

