% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dodaj_info.R
\name{dodaj_info}
\alias{dodaj_info}
\title{Funkcja dodaje nazwę eksponatu na mapę cnk.}
\usage{
dodaj_info(id, pelna_nazwa)
}
\arguments{
\item{id}{Numer id eksponatu cnk.}

\item{pelna_nazwa}{Pełna nazwa eksponatu cnk o identyfikatorze id.}
}
\description{
Nanosi id oraz nazwę eksponatu na mapkę cnk.
}

