% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generuj.sciezke.R
\name{generuj.sciezke}
\alias{generuj.sciezke}
\title{Funkcja generuje ścieżkę odwiedzanych eksponatów cnk.}
\usage{
generuj.sciezke(first, l, mat)
}
\arguments{
\item{first}{Numer id pierwszego eksponatu w ścieżce.}

\item{l}{Pożądana długość generowanej ścieżki.}

\item{mat}{Macierz prawdopodobieństw przejścia}
}
\value{
sciezka Ścieżka dlugości "l" od eksponatu "first".
}
\description{
Generuje ścieżkę odwiedzanych eksponatów cnk na podstawie macierzy przejścia
 w kolejnych krokach.
}

