% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapka1.R
\name{mapka1}
\alias{mapka1}
\title{Funkcja rysuje mapę cnk}
\usage{
mapka1(wybrany, bliskie, eksponaty, wspX, wspY, img, srodowisko, usr)
}
\arguments{
\item{wybrany}{Numer id wybranego eksponatu.}

\item{bliskie}{Numery id punktów bliskich do wybranego.}

\item{eksponaty}{Numery id wszystkich ekponatów.}

\item{wspX}{Współrzędne x wszystkich eksponatów.}

\item{wspY}{Współrzędne y wszystkich eksponatów.}

\item{img}{Obrazek w postaci obiektu array wczytany funkcją readPNG z pakietu png.}

\item{srodowisko}{Środowisko, w którym przechowywane są zmienne globalne 
dla bieżącej sesji aplikacji.}

\item{usr}{Parametr usr zwracany przez funkcję par().}
}
\description{
Rysuje mapę cnk. Wybrany eksponat zaznacz kolorem niebieskim, bliskie eksponaty
kolorem czarnym, a pozostałe kolorem szarym.
}

