% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pstwa.przejscia.R
\name{pstwa.przejscia}
\alias{pstwa.przejscia}
\title{Funkcja obliczająca macierz prawdopodobieńst przejscia}
\usage{
pstwa.przejscia(baza, wyjscie = "pstwa.rda")
}
\arguments{
\item{baza}{Ścieżka do wejściowej bazy SQLitowej.}

\item{wyjscie}{Sciezka do pliku wyjściowego.}
}
\description{
Funkjca oblicza macierz prawdopodobieństw przejścia w jednym kroku z jednego
eksponatu do drugiego dla każdej pary eksponatów. Dane są wczytywane z bazy
SQLite znajdującej się w pliku baza i zapisywane są do pliku .rda o nazwie
wyjscie.
}

