% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reaguj.mapa.R
\name{reaguj.mapa}
\alias{reaguj.mapa}
\title{Funkcja rysująca aktualną mapę po kliknięciu na wybrany eksponat.}
\usage{
reaguj.mapa(eksponaty, klik_mapa, wspX, wspY, srodowisko, session, usr)
}
\arguments{
\item{eksponaty}{Eksponaty spośród których wybieramy następny.}

\item{klik_mapa}{Zdarzenie kliknięcia na mapę.}

\item{wspX}{Współrzędne x eksponatów cnk.}

\item{wspY}{Współrzędne y eksponatów cnk.}

\item{srodowisko}{Środowisko, w którym przechowywane są zmienne globalne 
dla bieżącej sesji aplikacji.}

\item{session}{Parametr z funkcji shinyServer}

\item{usr}{Parametr usr zwracany przez funkcję par().}
}
\description{
Po kliknięciu w wybrany eksponat rysuje zaktualizowaną mapę przejść
do kolejnych eksponatów cnk.
}

