% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reaguj.wykres.R
\name{reaguj.wykres}
\alias{reaguj.wykres}
\title{Funkcja rysuje zaktualizowany wykres prawdopodobieństw przejść.}
\usage{
reaguj.wykres(eksponaty, klik, session)
}
\arguments{
\item{eksponaty}{Eksponaty bliskie do eksponatu, na który klikniemy.}

\item{klik}{Zdarzenie kliknięcia na wykres}

\item{session}{Parametr z funkcji shinyServer}
}
\description{
Zmienia wybrany eksponat po kliknięciu na słupek eksponatu na wykresie słupkowym.
}

