% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rysuj.mape.R
\name{rysuj.mape}
\alias{rysuj.mape}
\title{Funkcja rysuje mapę cnk.}
\usage{
rysuj.mape(img, srodowisko)
}
\arguments{
\item{img}{Obrazek w postaci obiektu array wczytany funkcją readPNG z pakietu png.}

\item{srodowisko}{Środowisko, w którym przechowywane są zmienne globalne 
dla bieżącej sesji aplikacji.}
}
\description{
Rysuje mapę cnk (bez zaznaczonych eksponatów).
}

