% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wczytaj.dane.R
\name{wczytaj.dane}
\alias{wczytaj.dane}
\title{Funkcja wczytuje dane z logów.}
\usage{
wczytaj.dane(sciezka, baza)
}
\arguments{
\item{sciezka}{Ścieżka do pliku z logami cnk.}

\item{baza}{Baza, do której będą zapisywane dane.}
}
\description{
Wczytuje dane z logów cnk i zapisuje je do bazy danych RSQLite.
}

