% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wczytaj.nazwy.plikow.R
\name{wczytaj.nazwy.plikow}
\alias{wczytaj.nazwy.plikow}
\title{Funkcja wczytuje nazwy plikow cnk.}
\usage{
wczytaj.nazwy.plikow(folder)
}
\arguments{
\item{folder}{Ścieżka, w której szukamy logów cnk.}
}
\value{
pliki Nazwy plików z logami cnk.
}
\description{
Wczytuje wczytuje nazwy plików postaci cnk[liczba] znajdujących się
w katalogu folder.
}

