% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wspX.R
\docType{data}
\name{wspX}
\alias{wspX}
\title{Współrzędne x eksponatów cnk.}
\format{Wektor numeryczny o długości 59.}
\description{
Wektor zawierający współrzędne x eksponatów na mapce cnk
(kolejnosc współrzędnych odpowiada kolejności nazw wektora "eksponaty").
}

