% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wspY.R
\docType{data}
\name{wspY}
\alias{wspY}
\title{Współrzędne y eksponatów cnk.}
\format{Wektor numeryczny o długości 59.}
\description{
Wektor zawierający współrzędne y eksponatów na mapce cnk
(kolejnosc współrzędnych odpowiada kolejności nazw wektora "eksponaty").
}

