% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wykres.R
\name{wykres}
\alias{wykres}
\title{Funkcja rysuje wykres prawdopodobieństw przejścia dla danego eksponatu cnk.}
\usage{
wykres(eksponaty, wybranyEksponat)
}
\arguments{
\item{eksponaty}{Eksponaty bliskie do "wybranyEksponat".}

\item{wybranyEksponat}{Eksponat, dla którego chcemy zonaczyć wykres 
prawdopodobieństw przejścia do innych eksponatów cnk.}
}
\description{
Rysuje wykres słupkowy prawdopodobieństw przejścia z danego eksponatu do 
innych eksponatów cnk.
}

