% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zamien.uklad.R
\name{zamien.uklad}
\alias{zamien.uklad}
\title{Funkcja zamienia układ współrzędnych.}
\usage{
zamien.uklad(wspX, wspY, rozmiarX, rozmiarY, usr)
}
\arguments{
\item{wspX}{Współrzędne x punktów na mapie eksponatów.}

\item{wspY}{Współrzędne y punktów na mapie eksponatów.}

\item{rozmiarX}{Szerokość obrazu z pikselach.}

\item{rozmiarY}{Wysokość obrazu w pikselach.}

\item{usr}{Parametr usr zwracany przez funkcję par().}
}
\value{
list Lista zawierająca nowy ukłąd współrzędnych (współrzędne x i y).
}
\description{
Zamienia wspX i wspY z układu współrzędnych odpowiającego obrazkowi 
o rozmiarach rozmiarX x roazmiarY na układ wykresu R o danym parametrze usr 
zwracanym przez funkcję par.
}

