% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zaznacz.eksponat.R
\name{zaznacz.eksponat}
\alias{zaznacz.eksponat}
\title{Funkcja zaznacza eksponaty na mapce cnk.}
\usage{
zaznacz.eksponat(eks, wspX, wspY, cex, ...)
}
\arguments{
\item{eks}{Nazwa eksponatu.}

\item{wspX}{Współrzędne x eksponatów cnk na mapie cnk.}

\item{wspY}{Współrzędne y eksponatów cnk na mapie cnk.}
}
\description{
Zaznacza na mapie cnk punkt, w którym znajduje się eksponat cnk.
}

